refactor: split PromptDetailView.tsx into focused sub-components (#94)
Break down 1204-line monolithic PromptDetailView into 9 focused modules:
- promptDetail/constants.ts: types and constants
- promptDetail/evidence.ts: utility functions (buildInjectedEvidence, getLanguage, etc.)
- promptDetail/StatPill.tsx: stat pill component
- promptDetail/Section.tsx: collapsible section with animation
- promptDetail/LegendDot.tsx: legend dot item
- promptDetail/TokenRow.tsx: token breakdown row
- promptDetail/EvidenceGroup.tsx: injected evidence group
- promptDetail/BreakdownPopover.tsx: context breakdown popover
- promptDetail/FilePreviewOverlay.tsx: file preview overlay with syntax highlighting
- promptDetail/index.ts: barrel exports

Also adds vitest infrastructure (vitest.config.ts, test setup, 5 tests).
All tests pass with no changes.
