+++ src/components/dashboard/promptDetail/index.ts
@@ -1,0 +1,16 @@
+export { StatPill } from "./StatPill";
+export { Section } from "./Section";
+export { LegendDot } from "./LegendDot";
+export { TokenRow } from "./TokenRow";
+export { EvidenceGroup } from "./EvidenceGroup";
+export { BreakdownPopover } from "./BreakdownPopover";
+export { FilePreviewOverlay } from "./FilePreviewOverlay";
+export { buildInjectedEvidence, getLanguage, normalizeText, getFileName } from "./evidence";
+export type { EvidenceStatus, InjectedEvidenceItem, PromptDetailViewProps } from "./constants";
+export {
+  CONTINUATION_PROMPT_MARKER,
+  SESSION_SCAN_DEDUP_MS,
+  COMPACTION_DROP_RATIO,
+  MIN_COMPACTION_BASE_TOKENS,
+  DIRECT_FILE_ACTIONS,
+} from "./constants";
