+++ vitest.config.ts
@@ -1,0 +1,19 @@
+import { defineConfig } from "vitest/config";
+import react from "@vitejs/plugin-react";
+import path from "path";
+
+export default defineConfig({
+  plugins: [react()],
+  resolve: {
+    alias: {
+      "@": path.resolve(__dirname, "./src"),
+    },
+  },
+  test: {
+    environment: "jsdom",
+    include: ["src/**/__tests__/**/*.{test,spec}.{ts,tsx}"],
+    setupFiles: ["src/__tests__/setup.ts"],
+    globals: true,
+    css: false,
+  },
+});
